Issue #8: Fix Chrome crash when Media plugin installed.
There is a bug in Chrome 48 when trying to override MediaError:
1. It doesn't work using Ripple's current approach.
2. It crashes Chrome if you try to do it twice.

This occurs when using the Media plugin (Ripple sets MediaError once, then the Media plugin sets it). It also causes the crash people are seeing if they try to navigate pages in a Cordova app running in Ripple (it is set a second time on the outer window after navigating).

Which this is a Chrome bug, there is a workaround: provide a getter in Object.defineProperty() rather than specifying a value.

Note that this change also fixes issue #7 (Geolocation not working) - like MediaError, Ripple was not successfully overriding window.navigator in Chrome 48, but with the new approach it works.

This resolves #7
This resolves #8