Issue #9: Add support for Media.MessageChannel exec call.
Version 1.0.1 of the Media plugin added a new exec call (Media.MessageChannel) that is not handled by Ripple. As a result, the "I Haz Cheeseburger?!?!" dialog displays when using this plugin in RIpple.

This change adds a handler for that exec call (and uses the provided method for callbacks).

This resolves #9